requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
pandas==2.1.3
openpyxl==3.1.2
python-multipart==0.0.6
aiofiles==23.2.1
openai==1.3.7
python-dotenv==1.0.0
pydantic-settings==2.0.3
